Add comments, clean up code, refactor file structs into new file

+++ student-distrib/include/kernel/syscall.h
@@ -23,7 +23,7 @@ int32_t syscall_write(int32_t fd, const void *buf, int32_t nbytes);
 int32_t syscall_open(const uint8_t *filename);
 int32_t syscall_close(int32_t fd);
 int32_t syscall_execute(const int8_t *command);
-int32_t syscall_halt(uint8_t status);
+int32_t syscall_halt(uint32_t status);
 int32_t syscall_getargs(uint8_t *buf, int32_t nbytes);
 int32_t syscall_vidmap(uint8_t **screen_start);
 int32_t syscall_set_handler(int32_t signum, void *handler_address);
